Workaround for inactive Users
Issue #278
This will display them so they can be removed. If would be better to go through and fix every place where we look at users, but this gets us most of the way. The only way users get inactived is if we do it manually when we notice emails bouncing.